Initial import of templates from blueprint-sass.
Some deviations from blueprint's styles have to be rectified still.

The "tests" have moved to the examples directory. The
examples can be built by running "rake examples"

The blueprint directory is still a set of ready-to-use
css files. These are generated by running "rake blueprint".
The default example's stylesheets are currently used to
generate these default css files.

"rake tests" will run tests that verify that no sass
compilation errors occur for any of the template fixtures. If
they do, the output is captured for debugging.

There's still much to do, which I'm keeping track of in the TODO file.